refactor(discovery): separate runtime, artifacts, matching, and prompts
