+++ packages/discovery/src/core/model.ts
