+++ packages/discovery/src/index.ts
@@ -1,8 +1,10 @@
-export { Artifacts } from './artifacts.js';
-export type { ArtifactInput, HydeState, HydeDocumentState } from './artifacts.js';
-export { Discovery, DISCOVERY_MIN_SIMILARITY, validateDiscoveryMinSimilarity, buildDiscovererContext } from './discovery.js';
-export type { DiscoveryDeps, DiscoveryInput, DiscoveryState, PotentialIntentPair, SourceProfileData } from './discovery.js';
-export { MatchExplainer } from './explanation.js';
-export type { MatchExplainerLike, MatchExplainerInput, MatchExplainerResult, EvaluatorEntity } from './explanation.js';
-export { ModelClient } from './model.js';
-export type { Model, ModelRequest, DiscoveryData, CandidateSearch, IntentCandidate, SearchOptions, ArtifactStore, ArtifactCache, EmbeddingGenerator, MatchEvidence, Lens, HydeTargetCorpus, RunOptions, Logger, TraceEvent } from './types.js';
+export { Artifacts } from './artifacts/artifact.pipeline.js';
+export type { ArtifactInput, HydeState, HydeDocumentState } from './artifacts/artifact.state.js';
+export { Discovery } from './matching/discovery.pipeline.js';
+export { DISCOVERY_MIN_SIMILARITY, validateDiscoveryMinSimilarity } from './matching/discovery.constants.js';
+export { buildDiscovererContext } from './prompts/discovery.prompt.js';
+export type { DiscoveryDeps, DiscoveryInput, DiscoveryState, PotentialIntentPair, SourceProfileData } from './matching/discovery.state.js';
+export { MatchExplainer } from './matching/match.explainer.js';
+export type { MatchExplainerLike, MatchExplainerInput, MatchExplainerResult, EvaluatorEntity } from './matching/discovery.state.js';
+export { ModelClient } from './core/model.js';
+export type { Model, ModelRequest, DiscoveryData, CandidateSearch, IntentCandidate, SearchOptions, ArtifactStore, ArtifactCache, EmbeddingGenerator, MatchEvidence, Lens, HydeTargetCorpus, RunOptions, Logger, TraceEvent } from './core/types.js';
